+++ saleor/graphql/shipping/types.py
@@ -4,6 +4,7 @@ from graphene import relay
 from ...core.permissions import ShippingPermissions
 from ...core.weight import convert_weight_to_default_weight_unit
 from ...shipping import models
+from ..channel import ChannelQsContext
 from ..channel.dataloaders import ChannelByIdLoader
 from ..channel.types import (
     ChannelContext,
@@ -11,6 +12,7 @@ from ..channel.types import (
     ChannelContextTypeWithMetadata,
 )
 from ..core.connection import CountableDjangoObjectType
+from ..core.fields import ChannelContextFilterConnectionField
 from ..core.types import CountryDisplay, Money, MoneyRange
 from ..decorators import permission_required
 from ..meta.types import ObjectWithMetadata
@@ -23,6 +25,7 @@ from .dataloaders import (
     ShippingMethodChannelListingByShippingMethodIdLoader,
     ShippingMethodsByShippingZoneIdAndChannelSlugLoader,
     ShippingMethodsByShippingZoneIdLoader,
+    ZipCodeRulesByShippingMethodIdLoader,
 )
 from .enums import ShippingMethodTypeEnum
 
@@ -45,6 +48,20 @@ class ShippingMethodChannelListing(CountableDjangoObjectType):
         return ChannelByIdLoader(info.context).load(root.channel_id)
 
 
+class ShippingMethodZipCodeRule(CountableDjangoObjectType):
+    start = graphene.String(description="Start address range.")
+    end = graphene.String(description="End address range.")
+
+    class Meta:
+        description = "Represents shipping method zip code."
+        interfaces = [relay.Node]
+        model = models.ShippingMethodZipCodeRule
+        only_fields = [
+            "start",
+            "end",
+        ]
+
+
 class ShippingMethod(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
     type = ShippingMethodTypeEnum(description="Type of the shipping method.")
     translation = TranslationField(
@@ -65,6 +82,14 @@ class ShippingMethod(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
     minimum_order_price = graphene.Field(
         Money, description="The price of the cheapest variant (including discounts)."
     )
+    zip_code_rules = graphene.List(
+        ShippingMethodZipCodeRule,
+        description="Zip code exclude range of the shipping method.",
+    )
+    excluded_products = ChannelContextFilterConnectionField(
+        "saleor.graphql.product.types.products.Product",
+        description="List of excluded products for the shipping method.",
+    )
 
     class Meta:
         default_resolver = ChannelContextType.resolver_with_context
@@ -78,6 +103,8 @@ class ShippingMethod(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
             "id",
             "maximum_order_weight",
             "minimum_order_weight",
+            "maximum_delivery_days",
+            "minimum_delivery_days",
             "name",
         ]
 
@@ -136,6 +163,12 @@ class ShippingMethod(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
         return convert_weight_to_default_weight_unit(root.node.maximum_order_weight)
 
     @staticmethod
+    def resolve_zip_code_rules(
+        root: ChannelContext[models.ShippingMethod], info, **_kwargs
+    ):
+        return ZipCodeRulesByShippingMethodIdLoader(info.context).load(root.node.id)
+
+    @staticmethod
     def resolve_minimum_order_weight(
         root: ChannelContext[models.ShippingMethod], *_args
     ):
@@ -149,6 +182,13 @@ class ShippingMethod(ChannelContextTypeWithMetadata, CountableDjangoObjectType):
         return ShippingMethodChannelListingByShippingMethodIdLoader(info.context).load(
             root.node.id
         )
+
+    @staticmethod
+    @permission_required(ShippingPermissions.MANAGE_SHIPPING)
+    def resolve_excluded_products(
+        root: ChannelContext[models.ShippingMethod], _info, **_kwargs
+    ):
+        return ChannelQsContext(qs=root.node.excluded_products.all(), channel_slug=None)
 
 
 class ShippingZone(ChannelContextTypeWithMetadata, CountableDjangoObjectType):