+++ saleor/graphql/shop/types.py
@@ -12,21 +12,24 @@ from ...core.permissions import SitePermissions, get_permissions
 from ...core.utils import get_client_ip, get_country_by_ip
 from ...plugins.manager import get_plugins_manager
 from ...site import models as site_models
-from ..account.types import Address, StaffNotificationRecipient
+from ..account.types import Address, AddressInput, StaffNotificationRecipient
 from ..channel import ChannelContext
 from ..checkout.types import PaymentGateway
+from ..core.connection import CountableDjangoObjectType
 from ..core.enums import WeightUnitsEnum
 from ..core.types.common import CountryDisplay, LanguageDisplay, Permission
 from ..core.utils import str_to_enum
 from ..decorators import permission_required
 from ..menu.dataloaders import MenuByIdLoader
 from ..menu.types import Menu
+from ..shipping.types import ShippingMethod
 from ..translations.enums import LanguageCodeEnum
 from ..translations.fields import TranslationField
 from ..translations.resolvers import resolve_translation
 from ..translations.types import ShopTranslation
 from ..utils import format_permissions_for_display
 from .enums import AuthorizationKeyType
+from .resolvers import resolve_available_shipping_methods
 
 
 class Navigation(graphene.ObjectType):
@@ -64,6 +67,13 @@ class Geolocalization(graphene.ObjectType):
         description = "Represents customers's geolocalization data."
 
 
+class OrderSettings(CountableDjangoObjectType):
+    class Meta:
+        only_fields = ["automatically_confirm_all_new_orders"]
+        description = "Order related settings from site settings."
+        model = site_models.SiteSettings
+
+
 class Shop(graphene.ObjectType):
     available_payment_gateways = graphene.List(
         graphene.NonNull(PaymentGateway),
@@ -74,6 +84,23 @@ class Shop(graphene.ObjectType):
         ),
         description="List of available payment gateways.",
         required=True,
+    )
+    available_shipping_methods = graphene.List(
+        ShippingMethod,
+        channel=graphene.Argument(
+            graphene.String,
+            description="Slug of a channel for which the data should be returned.",
+            required=True,
+        ),
+        address=graphene.Argument(
+            AddressInput,
+            description=(
+                "Address for which available shipping methods should be returned."
+            ),
+            required=False,
+        ),
+        required=False,
+        description="Shipping methods that are available for the shop.",
     )
     geolocalization = graphene.Field(
         Geolocalization, description="Customer's geolocalization data."
@@ -169,6 +196,10 @@ class Shop(graphene.ObjectType):
     @staticmethod
     def resolve_available_payment_gateways(_, _info, currency: Optional[str] = None):
         return get_plugins_manager().list_payment_gateways(currency=currency)
+
+    @staticmethod
+    def resolve_available_shipping_methods(_, info, channel, address=None):
+        return resolve_available_shipping_methods(info, channel, address)
 
     @staticmethod
     @permission_required(SitePermissions.MANAGE_SETTINGS)