+++ CHANGELOG.md
@@ -13,8 +13,13 @@ All notable, unreleased changes to this project will be documented in this file.
 - Drop deprecated service accounts and webhooks API - #6431 by @maarcingebala
 - Add editorjs sanitizer - #6456 by @IKarbowiak
 - Add generic FileUpload mutation - #6470 by @IKarbowiak
+- Order confirmation backend - #6498 by @tomaszszymanski129
 - Multichannel MVP: Multicurrency - #6242 by @fowczarek @d-wysocki
 - Fix password reset request - #6351 by @Manfred-Madelaine-pro, Ambroise and Pierre
+- Refund products support - #6530 by @korycins
+- Add possibility to exclude products from shipping method - #6506 by @korycins
+- Add availableShippingMethods to the Shop type - #6551 by @IKarbowiak
+- Add delivery time to shipping method - #6564 by @IKarbowiak
 
 # 2.11.1